Experiment with two distributions
This is done because the hesslab images have very bright fiducials
embedded.  Using one distribution with a high maximum doesn't work well
because there are too many emitters of intermediate intensity in the
simulation.  So instead experiment with using two distributions one for
fiducials and one for the blinking emitters.